feat: Add AlertsCard component and AlertsModal for improved alert management; include new icons
